EASE/requirements.txt
pandas==2.0.1
scipy==1.8.0
hydra-core==1.3.2